extract menu-building helper from CollectionNav
